Changes to fix circular include
